src/main/java/br/usp/each/saeg/asm/defuse/ObjectField.java
/**
 * asm-defuse: asm powered by definitions/uses analysis
 * Copyright (c) 2013, 2020 Roberto Araujo (roberto.andrioli@gmail.com)
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions
 * are met:
 * 1. Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 * 2. Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in the
 *    documentation and/or other materials provided with the distribution.
 * 3. Neither the name of the copyright holders nor the names of its
 *    contributors may be used to endorse or promote products derived from
 *    this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
 * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
 * THE POSSIBILITY OF SUCH DAMAGE.
 */
package br.usp.each.saeg.asm.defuse;

import java.util.Collections;
import java.util.LinkedHashSet;
import java.util.Set;

import org.objectweb.asm.Type;

public final class ObjectField extends Field {

    public final Value value;

    public ObjectField(final String owner, final String name, final String desc, final Value value) {
        super(owner, name, desc);
        if (value.type.getSort() != Type.OBJECT) {
            throw new IllegalArgumentException("Invalid value type: " + value.type);
        }
        this.value = value;
    }

    @Override
    public Set<Variable> getVariables() {
        final Set<Variable> values = new LinkedHashSet<Variable>();
        values.addAll(value.getVariables());

        final Value root = getRoot();
        if (root instanceof Local || root instanceof StaticField) {
            values.add(this);
        }

        return Collections.unmodifiableSet(values);
    }

    public Value getRoot() {
        Value root = value;
        while (root instanceof ObjectField) {
            root = ObjectField.class.cast(root).value;
        }
        return root;
    }

    @Override
    public String toString() {
        return String.format("%s.%s.%s", value, owner.replace("/", "."), name);
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = super.hashCode();
        result = prime * result + value.hashCode();
        return result;
    }

    @Override
    public boolean equals(final Object obj) {
        if (!super.equals(obj))
            return false;

        final ObjectField other = (ObjectField) obj;

        if (!value.equals(other.value))
            return false;

        return true;
    }

}
